+++ errors.go
@@ -99,8 +99,3 @@ func IsRatelimiterAcquireError(err error) bool {
 func ToRatelimiterAcquireError(err error) ratelimiter.AcquireError {
 	return ratelimiterErrors.ToAcquireError(err)
 }
-
-// PrintErrorWithoutStack removed stacktrace records from error string
-func PrintErrorWithoutStack(err error) string {
-	return xerrors.PrintWithoutStack(err)
-}